packages/server/pyproject.toml
[tool.poetry]
name = "insighthub-server"
version = "2.0.0"
description = "Simplified RAG System with Vector and Graph implementations"
authors = ["Nate Stott"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-cors = "^6.0.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
qdrant-client = "^1.7.0"
sentence-transformers = "^2.2.2"
requests = "^2.31.0"
pyjwt = "^2.8.0"
werkzeug = "^3.0.0"
PyPDF2 = "^3.0.0"
flask-socketio = "^5.5.1"
python-socketio = "^5.15.0"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
bcrypt = "^5.0.0"
pypdf = "^6.4.0"
python-json-logger = "^4.0.0"
redis = "^7.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"
ruff = "^0.6.0"
types-requests = "^2.31.0"
types-flask-cors = "^6.0.0.20250809"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

# isort configuration
[tool.isort]
profile = "black"
line_length = 100

# Mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py311"

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
